device: move tile type independent array functions to a new file
